Move swagger_schema DSL macro to top level PhoenixSwagger module
 - Most important macros now all exist at the top level module
