+++ lib/phoenix_swagger/schema.ex
@@ -82,49 +82,4 @@ defmodule PhoenixSwagger.Schema do
 
     %{model | properties: properties, required: required}
   end
-
-  @doc """
-  This macro accepts a block where the schema is defined by listing the properties using a DSL.
-  Defining a property for the schema takes 2 required parameters and one optional parameter.
-
-  The first parameter is the name of the property.
-  The second parameter is the property type, either a simple type like :string, or a %Schema{} struct.
-  The third parameter is the description of the property.
-  The fourth optional parameter is an optional list of parameters that is passed to the PhoenixSwagger.Schema.
-
-  All these values are taken and placed into a PhoenixSwagger.Schema struct.
-
-  The PhoenixSwagger.Schema struct is returned from a `schema/0` function that is provided by this macro.
-
-  Example:
-    swagger_schema do
-      full_name, :string, "Full name"
-      title, :string, "Title", required: true
-      genre, :string, "Genre", enum: [:scifi, :horror, :drama, :comedy]
-      birthday, :datetime, "Birthday", format: :datetime
-    end
-  """
-  defmacro swagger_schema(properties_block) do
-    properties = case properties_block do
-      [do: {:__block__, _, info}] -> info
-      [do: info] -> [info]
-    end
-
-    model = %Schema{type: :object}
-
-    body =
-      properties
-      |> Enum.map(fn {name, line, args} -> {:property, line, [name | args]} end)
-      |> Enum.reduce(Macro.escape(model), fn next, pipeline ->
-          quote do
-            unquote(pipeline) |> unquote(next)
-          end
-        end)
-
-    quote do
-      def schema do
-        unquote(body)
-      end
-    end
-  end
 end
